Code of GUI for Laser-Cutting Microscope
gui.py is the main code. NPILaser.py and ESP301.py are classes that handle the communication to the laser and the stage respectively. GuiHelper.py and ClickableCameraLabel.py are files that contain helper classes for the gui.py. amcam.py and amcam.dll are the files for the communication with the camera.